renamed some files for clarity, negligible change to API, bump version to 0.8.2
