+++ docs/source/conf.py
@@ -4,7 +4,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../.
 
 project = 'IceGraph'
 author = 'Taylor St Jean'
-release = '0.8.1'
+release = '0.8.2'
 
 html_theme = "sphinx_rtd_theme"
 master_doc = 'index'
