+++ icegraph/console/console.py
