+++ icegraph/data/base/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import IGData
+from .data import IGData
 
 IGData.__module__ = __name__
