+++ icegraph/data/converters/base/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import IGConverter
+from .base import IGConverter
 
 __all__ = ["IGConverter"]
