+++ icegraph/data/readers/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import LMDBDatasetShardReader, LMDBReader
+from .readers import LMDBDatasetShardReader, LMDBReader
 
 LMDBDatasetShardReader.__module__ = __name__
 LMDBReader.__module__ = __name__
