+++ icegraph/data/writers/base/base.py
