+++ icegraph/geometry/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import Detector
+from .detector import Detector
 
 Detector.__module__ = __name__
