+++ icegraph/geometry/detector.py
