+++ icegraph/pathutils/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import PathResolver, PathValidator
+from icegraph.utils.pathutils import PathResolver, PathValidator
 
 PathResolver.__module__ = __name__
 PathValidator.__module__ = __name__
