+++ icegraph/renderer/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import CDFPlot, PDFPlot, ChargeDistributionPlot, ParityPlot
+from .plots import CDFPlot, PDFPlot, ChargeDistributionPlot, ParityPlot
 
 CDFPlot.__module__ = __name__
 PDFPlot.__module__ = __name__
