+++ icegraph/renderer/base/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import IGBasicPlot, IGDistributionPlot
+from .base import IGBasicPlot, IGDistributionPlot
 
 IGBasicPlot.__module__ = __name__
 IGDistributionPlot.__module__ = __name__
