+++ icegraph/trainer/callbacks/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import ExportCallback, ConsoleCallback, TensorBoardCallback, RegressionMetricsCallback
+from .callbacks import ExportCallback, ConsoleCallback, TensorBoardCallback, RegressionMetricsCallback
 from .normalizers import MinMaxNormalizer
 
 ExportCallback.__module__ = __name__
