+++ icegraph/trainer/callbacks/base/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import Callback, Normalizer
+from .base import Callback, Normalizer
 
 Callback.__module__ = __name__
 Normalizer.__module__ = __name__
