+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = icegraph
-version = 0.8.1
+version = 0.8.2
 
 [options]
 packages = find:
